+++ cart/models.py
@@ -16,7 +16,7 @@ class Cart(models.Model):
 class CartItem(models.Model):
 	product = models.ForeignKey(Product, on_delete=models.CASCADE) 
 	cart = models.ForeignKey(Cart, on_delete=models.CASCADE)
-	quantity = models.IntegerField(null=True)
+	quantity = models.BigIntegerField(null=True)
 	active = models.BooleanField(default=True)
 
 	class Meta:
